+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Upcoming - 3.0 - WebAssembly
 ---
-- jsNet version 1.5 feature parity
-- Global bug fixes
-- Removed babel transpilation (it's 2018)
+#### WebAssembly
+- Complete, rewritten, WebAssembly version of jsNet
+
+#### Global
+- Bug fixes
 - Removed default configuration values: l1, l2
+- Added layer specific activation function config for FC layers, and ability to turn it off
+
+#### JavaScript
+- Removed babel transpilation (it's 2018)
 
 # 2.1.0 - Optimizations
 ---